requirements.txt
Django==3.1
pycodestyle==2.5.0
pep8==1.7.1
djangorestframework==3.12.1
markdown==3.3.3
django-filter==2.4.0
